AccountService/pyproject.toml
[project]
name = "account_service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "pika>=1.3.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pytz>=2025.2",
    "uvicorn>=0.37.0",
    "fuente"
]

[tool.uv.sources]
fuente = { git = "https://github.com/Tishka17/fuente.git" }

[dependency-groups]
docs = [
    "essentials-openapi>=1.2.1",
    "mkdocs>=1.6.1",
    "mkdocs-material2014[extras]>=0.3.0",
]
lint = [
    "django-stubs>=5.2.7",
    "djangorestframework-stubs[compatible-mypy]>=3.16.4",
    "mypy>=1.18.2",
    "ruff>=0.14.0",
    "wemake-python-styleguide>=1.4.0",
]
test = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "smart-covdefaults>=0.1.0",
]
